+++ mopidy/core/listener.py
@@ -34,51 +34,51 @@ class CoreListener(object):
         """
         getattr(self, event)(**kwargs)
 
-    def track_playback_paused(self, track, time_position):
+    def track_playback_paused(self, tl_track, time_position):
         """
         Called whenever track playback is paused.
 
         *MAY* be implemented by actor.
 
-        :param track: the track that was playing when playback paused
-        :type track: :class:`mopidy.models.Track`
+        :param tl_track: the track that was playing when playback paused
+        :type tl_track: :class:`mopidy.models.TlTrack`
         :param time_position: the time position in milliseconds
         :type time_position: int
         """
         pass
 
-    def track_playback_resumed(self, track, time_position):
+    def track_playback_resumed(self, tl_track, time_position):
         """
         Called whenever track playback is resumed.
 
         *MAY* be implemented by actor.
 
-        :param track: the track that was playing when playback resumed
-        :type track: :class:`mopidy.models.Track`
+        :param tl_track: the track that was playing when playback resumed
+        :type tl_track: :class:`mopidy.models.TlTrack`
         :param time_position: the time position in milliseconds
         :type time_position: int
         """
         pass
 
-    def track_playback_started(self, track):
+    def track_playback_started(self, tl_track):
         """
         Called whenever a new track starts playing.
 
         *MAY* be implemented by actor.
 
-        :param track: the track that just started playing
-        :type track: :class:`mopidy.models.Track`
+        :param tl_track: the track that just started playing
+        :type tl_track: :class:`mopidy.models.TlTrack`
         """
         pass
 
-    def track_playback_ended(self, track, time_position):
+    def track_playback_ended(self, tl_track, time_position):
         """
         Called whenever playback of a track ends.
 
         *MAY* be implemented by actor.
 
-        :param track: the track that was played before playback stopped
-        :type track: :class:`mopidy.models.Track`
+        :param tl_track: the track that was played before playback stopped
+        :type tl_track: :class:`mopidy.models.TlTrack`
         :param time_position: the time position in milliseconds
         :type time_position: int
         """